+++ _example/fiber/main.go
@@ -6,8 +6,7 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/gofiber/adaptor/v2"
-	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v3"
 	"github.com/soheilhy/cmux"
 
 	"github.com/arl/statsviz"
@@ -29,13 +28,15 @@ func main() {
 
 	// Fiber instance
 	app := fiber.New()
-	app.Get("/", func(c *fiber.Ctx) error {
-		return c.SendString("Hello, World 👋!")
-	})
 
-	// statsviz
-	app.Use("/debug/statsviz", adaptor.HTTPHandler(statsviz.Index))
-	ws.HandleFunc("/debug/statsviz/ws", statsviz.Ws)
+	// Create statsviz server.
+	srv, err := statsviz.NewServer()
+	if err != nil {
+		panic(err)
+	}
+
+	app.Use("/debug/statsviz/", srv.Index())
+	ws.HandleFunc("/debug/statsviz/ws", srv.Ws())
 
 	fmt.Println("Point your browser to http://localhost:8093/debug/statsviz/")